requirements_new.txt
bleach!=5.0.0
cmake==3.30.3
dlib==19.24.6
gdown==5.2.0
importlib-metadata==8.4.0
importlib-resources==6.4.5
imutils==0.5.4
mediapipe==0.9.3
oauthlib>=3.0.0
opencv-python==4.10.0.84
pandas==2.0.3
pyyaml==6.0.2
requests==2.32.3
scikit-learn==1.3.2
timm==1.0.9
torch==2.0.0 
torchaudio==2.0.1 
torchvision==0.15.1
tqdm==4.66.5

